# examples/bookmarks/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bookmarks",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/bookmarks/{id}/archive": {
      "post": {
        "tags": [
          "Bookmarks"
        ],
        "operationId": "api.bookmarks.archive",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/InvalidValues"
          },
          "401": {
            "description": "The API token is invalid, expired or revoked. The request isn't logged in: it has no valid API token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks an ability the operation needs: bookmarks:write.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": [
              "bookmarks:write"
            ]
          }
        ]
      }
    },
...
    "/api/v1/login": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "operationId": "api.login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidValues"
          },
          "401": {
            "description": "The API token is invalid, expired or revoked.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": []
      }
    },
...
    "/api/v1/logout": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "operationId": "api.logout",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "The API token is invalid, expired or revoked. The request isn't logged in: it has no valid API token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          },
          "device_name": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "two_factor": {
            "type": "boolean"
          },
          "challenge": {
            "type": "string"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "Problem details (RFC 9457).",
        "properties": {
          "type": {
            "type": "string",
            "description": "\"about:blank\": the status says what happened."
          },
          "title": {
            "type": "string",
            "description": "The status's title, in the request's language."
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status."
          },
          "detail": {
            "type": "string",
            "description": "What went wrong, for people."
          },
          "errors": {
            "type": "object",
            "description": "A message per field of the request that has a problem.",
            "additionalProperties": {
              "type": "string"
            }
          },
          "request_id": {
            "type": "string",
            "description": "The request's ID, as in the logs."
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "email_verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "email_verified_at",
          "created_at"
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "An error: problem details.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "InvalidInput": {
        "description": "The input isn't valid: problem details with an error per field.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "InvalidValues": {
        "description": "The request's values can't be read: problem details with an error per field.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API token, sent as \"Authorization: Bearer <token>\"."
      }
    }
  }
}